pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sample_agent"
version = "0.1.0"
description = "A sample agent implementation using OpenAI Agents SDK"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.68.0",
    "openai-agents>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.26.0",
    "uvicorn>=0.15.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/yourusername/sample_agent"
Documentation = "https://github.com/yourusername/sample_agent#readme"
Repository = "https://github.com/yourusername/sample_agent.git"

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "black>=21.9b0",
    "isort>=5.9.3",
    "flake8>=3.9.2",
]

[project.scripts]
sample-agent = "sample_agent.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sample_agent"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q"
